+++ pyvium/__init__.py
@@ -2,7 +2,7 @@ from .core import Core
 from .pyvium import (ChannelStatus, DEVICE_STATUS_LABELS, Pyvium, PyviumChannel,
                      PyviumInstance, device_status_label)
 from .tools import (DataPoint, ImpedancePoint, IndexEntry, MeasurementIndex,
-                    MeasurementInfo, MeasurementPart, MeasurementReader, Tools,
-                    UnsupportedDatabaseVersionError)
+                    MeasurementInfo, MeasurementPart, MeasurementPartSummary,
+                    MeasurementReader, Tools, UnsupportedDatabaseVersionError)
 from .instance_manager import (DiscoveryReport, IviumsoftInstanceManager,
                                ManagedInstance, UntrackedProcess)